Update README.md
This Looks Way more Inviting! Hehehe :) This is just a quick first Draft to showcase the possibilities. Needs more work. Also all the Screenshots should be uploaded into the github repo and the  pictures link urls in the readme should be change to point to this new github location... Peace!